fixed BeaconSetter drawing

+++ app/src/main/java/com/arielu/shopper/demo/DrawMapActivity.java
@@ -22,7 +22,7 @@ public class DrawMapActivity extends AppCompatActivity {
     private Line line;
     private LineView mlineView;
     private Intent intent;
-    private int width,height;
+    private float width,height;
 
 
     @Override
@@ -35,8 +35,8 @@ public class DrawMapActivity extends AppCompatActivity {
         mlineView = new LineView(this);
         mlineView.setBackgroundResource(R.drawable.map);
         setContentView(mlineView);
-        width = mlineView.getBackground().getIntrinsicWidth();
-        height = mlineView.getBackground().getIntrinsicHeight();
+        width = (float) mlineView.getBackground().getIntrinsicWidth();
+        height = (float) mlineView.getBackground().getIntrinsicHeight();
         mlineView.setFixWidth(width);
         mlineView.setFixHeight(height);
         fixPaths(sortedTracker);
